修改告警

+++ frameworks/core/components_ng/base/lazy_compose_adapter.cpp
@@ -48,7 +48,8 @@ void LazyComposeAdapter::SetActiveRange(int32_t start, int32_t end)
 
 void LazyComposeAdapter::OnDataChange(int32_t changeIndex)
 {
-    items_.RemoveIf([idx = static_cast<int32_t>(changeIndex)](const uint32_t& k, const auto& _) { return k >= idx; });
+    items_.RemoveIf([idx = static_cast<int32_t> (changeIndex)](
+                        const uint32_t& k, const auto& _) { return static_cast<int32_t>(k) >= idx; });
 }
 
 void LazyComposeAdapter::SetCallbacks(CreateItemCb create, UpdateRangeCb update)
